organisation page is almost finished but could not match the type totally being sent from server

+++ lib/src/data/model/offers_details_model.dart
@@ -4,7 +4,7 @@ class OffersDetailsModel {
   String? user;
   String? name;
   int? cost;
-  int? discount;
+  double? discount;
   int? qty;
   int? measurement;
   String? image;
